requirements.txt
PyYAML>=3.10
Cython>=0.29.21
crc32c
crcmod
cos-python-sdk-v5
numpy
yacs
scipy
scikit-learn
protobuf
pyautomaxprocs
pytorch_lightning
transformers==4.10.2
torchsnooper
albumentations
gpustat
jieba
timm==0.3.2
yacs
prefetch_generator
torchmetrics==0.5.1